# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsAggregator Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/preference": {
            "get": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Display specified user's preference",
                "operationId": "e936ff3f8a56b93995255c1db15d0712",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "preference": {
                                            "$ref": "#/components/schemas/UserPreference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Could not process the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Request failed, could not process your request at the moment, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Preferences"
                ],
                "summary": "Store user's  preference",
                "operationId": "d3b6e07e54aa40cb772100b0b08f3854",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserPreferenceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preference updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preference updated"
                                        },
                                        "preference": {
                                            "$ref": "#/components/schemas/UserPreference"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Request is empty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Request is empty"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Could not process the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Could not process your request at the moment, please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreUserPreferenceRequest": {
                "title": "StoreUserPreferenceRequest",
                "description": "Store User Preference Request body",
                "properties": {
                    "category": {
                        "description": "User's preferred category",
                        "type": "string",
                        "example": "sports"
                    },
                    "author": {
                        "description": "User's preferred author",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "source": {
                        "description": "User's preferred source",
                        "type": "string",
                        "example": "example.com"
                    }
                },
                "type": "object"
            },
...
            "UserPreference": {
                "title": "UserPreference",
                "description": "UserPreference model",
                "xml": {
                    "name": "UserPreference"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}